fix: --split-size was 32 bit, so asking for 4 GiB turned splitting off
A split size is a byte count and 4 GiB is an ordinary one, but it was held in a
uint32_t and truncated as it was parsed. Measured on a 421 MB source, where
every size below should produce a single part:

  --split-size=200mb    3 parts of ~209 MB    correct, and the control
  --split-size=4500mb   3 parts of ~209 MB    4,500,000,000 minus 2^32 leaves
                                              205,032,704, so it split at 205 MB
  --split-size=4gib     1 part, and NOT        4 GiB is exactly 2^32, so the size
                        named .split.1         became 0 and splitting was quietly
                                               switched off altogether

Widening the size alone would not have been enough. The running comparison
multiplied a uint32_t packet count by an int frame size, so the product was
worked out in 32 bit arithmetic and wrapped once a part passed 4 GB. Any split
size above 4 GB would then never be reached, because the figure it is compared
against starts counting from zero again at that point. Both are 64 bit now.

That half is proven by asking for 4.4 GB parts, just past the wrap, from 4.7 GB
of joined input: two parts, the first of 4,405,683,744 bytes. A part that size
is past 4,294,967,296 and could not have formed before.

A size that works out at zero or below is now refused rather than silently
turning splitting off, which is exactly what asking for 4 GiB used to do.

The interface needs no change. Its own control already accepts up to 99,999.99
of any unit and writes the value out as text, so it could always ask for sizes
the engine then truncated. Its exact wording, "4.500GB" and "5.000GiB", is
covered by the tests.

Small sizes are untouched: 200mb and 0.200GB still give the same three parts,
byte for byte.

+++ tsMuxer/tsMuxer.h
@@ -63,7 +63,7 @@ class TSMuxer final : public AbstractMuxer
     const std::vector<uint32_t>& getMuxedPacketCnt() { return m_muxedPacketCnt; }
     [[nodiscard]] size_t splitFileCnt() const { return m_fileNames.size(); }
     void setSplitDuration(const int64_t value) { m_splitDuration = value; }
-    void setSplitSize(const uint32_t value) { m_splitSize = value; }
+    void setSplitSize(const int64_t value) { m_splitSize = value; }
     void parseMuxOpt(const std::string& opts) override;
 
     void setFileName(const std::string& fileName, FileFactory* fileFactory) override;
@@ -135,7 +135,9 @@ class TSMuxer final : public AbstractMuxer
     int m_curFileNum;
     bool m_bluRayMode;
     bool m_hdmvDescriptors;
-    uint32_t m_splitSize;
+    // 64 bit: a split size is a byte count and 4 GiB is a perfectly ordinary one. As a uint32_t it
+    // wrapped, so 4 GiB became 0 and turned splitting off, and 4.5 GB became 205 MB.
+    int64_t m_splitSize;
     int64_t m_splitDuration;
 
     bool m_useNewStyleAudioPES;
